fix(agent-session): create worker socket directory
Ensure detached POSIX workers create and validate their private socket parent before binding, including source-checkout qualification paths.
